+++ src/main/java/ru/yandex/practicum/filmorate/model/Genre.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Genre {
-    private long id;
+    private int id;
     private String name;
+
 }
