+++ src/main/java/ru/yandex/practicum/filmorate/service/GenresService.java
@@ -8,6 +8,6 @@ public interface GenresService {
     List<Genre> getAll();
 
     // объект содержащий жанр с идентификатором id
-    Genre getByID(Long id);
+    Genre getByID(int id);
 
 }
